Restructure checks
