+++ qc_openmaterial3d/checks/__init__.py
@@ -4,7 +4,6 @@
 # Public License, v. 2.0. If a copy of the MPL was not distributed
 # with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from . import schema_checker as schema_checker
-from . import basic_checker as basic_checker
+from . import xom_general_checker as xom_general_checker
 from . import models as models
 from . import utils as utils
