--- qc_openmaterial3d/checks/basic_checker/__init__.py
@@ -1,8 +1,0 @@
-# SPDX-License-Identifier: MPL-2.0
-# Copyright 2024, ASAM e.V.
-# This Source Code Form is subject to the terms of the Mozilla
-# Public License, v. 2.0. If a copy of the MPL was not distributed
-# with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
-
-from . import valid_json_document as valid_json_document
-from . import version_is_defined as version_is_defined
