+++ qc_openmaterial3d/checks/xom_general_checker/__init__.py
@@ -4,4 +4,6 @@
 # Public License, v. 2.0. If a copy of the MPL was not distributed
 # with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
-from . import valid_schema as valid_schema
+from . import general_valid_json_document as valid_json_document
+from . import general_version_is_defined as version_is_defined
+from . import general_valid_schema as valid_schema
