+++ qc_openmaterial3d/checks/xom_general_checker/general_version_is_defined.py
@@ -13,14 +13,14 @@ from qc_baselib import IssueSeverity
 from qc_openmaterial3d import constants
 from qc_openmaterial3d.checks import models
 
-from qc_openmaterial3d.checks.basic_checker import (
-    valid_json_document,
+from qc_openmaterial3d.checks.xom_general_checker import (
+    general_valid_json_document,
 )
 
 CHECKER_ID = "check_asam.net:xom:1.0.0:general.version_is_defined"
 CHECKER_DESCRIPTION = "The metadata of the file must contain an openMaterialVersion field."
 CHECKER_PRECONDITIONS = {
-    valid_json_document.CHECKER_ID
+    general_valid_json_document.CHECKER_ID
 }
 RULE_UID = "asam.net:xom:1.0.0:general.version_is_defined"
