+++ tests/test_xom_general_checks.py
@@ -7,7 +7,7 @@
 import os
 import test_utils
 from qc_baselib import Result, IssueSeverity, StatusType
-from qc_openmaterial3d.checks import basic_checker
+from qc_openmaterial3d.checks import xom_general_checker
 
 
 def test_valid_json_document_positive(monkeypatch,) -> None:
@@ -23,7 +23,7 @@ def test_valid_json_document_positive(monkeypatch,) -> None:
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-            result.get_checker_status(basic_checker.valid_json_document.CHECKER_ID)
+            result.get_checker_status(xom_general_checker.valid_json_document.CHECKER_ID)
             == StatusType.COMPLETED
     )
     assert (
@@ -47,7 +47,7 @@ def test_valid_json_document_negative(monkeypatch,) -> None:
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-            result.get_checker_status(basic_checker.valid_json_document.CHECKER_ID)
+            result.get_checker_status(xom_general_checker.valid_json_document.CHECKER_ID)
             == StatusType.COMPLETED
     )
 
@@ -72,7 +72,7 @@ def test_version_is_defined_positive(monkeypatch,) -> None:
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-            result.get_checker_status(basic_checker.version_is_defined.CHECKER_ID)
+            result.get_checker_status(xom_general_checker.version_is_defined.CHECKER_ID)
             == StatusType.COMPLETED
     )
 
@@ -97,7 +97,7 @@ def test_version_is_defined_missing_attr(monkeypatch,) -> None:
     result.load_from_file(test_utils.REPORT_FILE_PATH)
 
     assert (
-            result.get_checker_status(basic_checker.version_is_defined.CHECKER_ID)
+            result.get_checker_status(xom_general_checker.version_is_defined.CHECKER_ID)
             == StatusType.COMPLETED
     )
 
@@ -106,4 +106,54 @@ def test_version_is_defined_missing_attr(monkeypatch,) -> None:
     )
     assert len(json_doc_issues) == 1
     assert json_doc_issues[0].level == IssueSeverity.ERROR
+    test_utils.cleanup_files()
+
+def test_valid_schema_positive(
+    monkeypatch,
+) -> None:
+    base_path = "tests/data/valid_schema/"
+    target_file_name = "json.valid_schema.positive.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_general_checker.valid_schema.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+
+    assert (
+        len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:general.valid_schema")) == 0
+    )
+
+    test_utils.cleanup_files()
+
+
+def test_valid_schema_negative(
+    monkeypatch,
+) -> None:
+    base_path = "tests/data/valid_schema/"
+    target_file_name = "json.valid_schema.negative.xoma"
+    target_file_path = os.path.join(base_path, target_file_name)
+
+    test_utils.create_test_config(target_file_path)
+
+    test_utils.launch_main(monkeypatch)
+
+    result = Result()
+    result.load_from_file(test_utils.REPORT_FILE_PATH)
+
+    assert (
+        result.get_checker_status(xom_general_checker.valid_schema.CHECKER_ID)
+        == StatusType.COMPLETED
+    )
+
+    assert (
+        len(result.get_issues_by_rule_uid("asam.net:xom:1.0.0:general.valid_schema")) == 1
+    )
     test_utils.cleanup_files()
